fix: improve api schema
