+++ packages/redis/README.md
@@ -1,0 +1,1 @@
+# Redis cache
