+++ packages/shared/README.md
@@ -1,0 +1,1 @@
+# Shared package
